pyproject.toml
[project]
name = "rag-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.2",
    "fastapi>=0.128.0",
    "openai>=2.16.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.46",
    "torch>=2.10.0",
    "transformers>=5.0.0",
    "uvicorn>=0.40.0",
]
